+++ mindsdb/integrations/handlers/neuralforecast_handler/requirements.txt
@@ -1,1 +1,2 @@
 neuralforecast>=1.4.0, <2.0
+ray>=2.0, <3.0
